+++ src/main.rs
@@ -292,6 +292,14 @@ pub struct Args {
     #[arg(long, help_heading = "Execution")]
     timeout_multiplier: Option<f64>,
 
+    /// Maximum run time for cargo build command, in seconds.
+    #[arg(long, help_heading = "Execution")]
+    build_timeout: Option<f64>,
+
+    /// Build timeout multiplier (relative to base build time).
+    #[arg(long, help_heading = "Execution")]
+    build_timeout_multiplier: Option<f64>,
+
     /// Print mutations that failed to check or build.
     #[arg(long, short = 'V', help_heading = "Output")]
     unviable: bool,