+++ testdata/hang_when_mutated/src/lib.rs
@@ -8,6 +8,18 @@ use std::time::{Duration, Instant};
 
 static TRIGGER: AtomicBool = AtomicBool::new(false);
 
+const fn should_stop_const() -> bool {
+    true
+}
+
+/// If `should_stop_const` is mutated to return false, then this const block
+/// will hang and block compilation.
+pub const VAL: i32 = loop {
+    if should_stop_const() {
+        break 1;
+    }
+};
+
 /// If mutated to return false, the program will spin forever.
 fn should_stop() -> bool {
     if TRIGGER.load(Ordering::Relaxed) {
@@ -43,5 +55,10 @@ mod test {
         // Should do two passes: first the trigger is false but gets set,
         // then the trigger is true and the loop terminates.
         assert_eq!(super::controlled_loop(), 2);
+    }
+
+    #[test]
+    fn val_is_correct() {
+        assert_eq!(super::VAL, 1);
     }
 }